+++ apps/react/src/app/app.tsx
@@ -16,6 +16,7 @@ import * as LitJsSdk_misc from 'dist/packages/misc';
 import * as LitJsSdk_miscBrowser from 'dist/packages/misc-browser';
 import * as LitJsSdk_nacl from 'dist/packages/nacl';
 import * as LitJsSdk_pkpBase from 'dist/packages/pkp-base';
+import * as LitJsSdk_pkpClient from 'dist/packages/pkp-client';
 import * as LitJsSdk_pkpCosmos from 'dist/packages/pkp-cosmos';
 import * as LitJsSdk_pkpEthers from 'dist/packages/pkp-ethers';
 import * as LitJsSdk_types from 'dist/packages/types';
@@ -38,6 +39,7 @@ declare global {
     LitJsSdk_miscBrowser: any;
     LitJsSdk_nacl: any;
     LitJsSdk_pkpBase: any;
+    LitJsSdk_pkpClient: any;
     LitJsSdk_pkpCosmos: any;
     LitJsSdk_pkpEthers: any;
     LitJsSdk_types: any;
@@ -1135,6 +1137,78 @@ export function App() {
 
     
     
+        if(typeof LitJsSdk_pkpClient === 'undefined') {
+            console.error("LitJsSdk_pkpClient:", LitJsSdk_pkpClient);
+        }else{
+            console.warn("LitJsSdk_pkpClient:", LitJsSdk_pkpClient);
+            window.LitJsSdk_pkpClient = LitJsSdk_pkpClient;
+        }
+        window.addEventListener('load', function() {
+
+            var root = document.getElementById('root');
+            var result = document.getElementById('result');
+            var entries = Object.entries(LitJsSdk_pkpClient);
+            var lis = entries.map(([key, value]) => `
+            <li>
+                <div id="LitJsSdk_pkpClient_${key}" class="key" onClick="(async (e) => {
+                    var fn = LitJsSdk_pkpClient['${key}'];
+                    var fnType = typeof fn;
+                    console.warn('[${key}] is type of [' + fnType + ']');
+
+                    if ( fnType === 'string' ) return;
+
+                    if( fnType === 'function' ){
+                        try{
+                            console.log('params:', globalThis.params);
+
+                            var res;
+                            try{
+                                res = new fn(globalThis.params);
+                            }catch{
+                                res = await fn(globalThis.params);
+                            }
+                            window.output = res;
+                            res = JSON.stringify(res, null, 2);
+                            result.innerText = res;
+                            console.log(res);
+                        }catch(e){
+                            console.error('Please set the [params] variable in the console then click again');
+                            console.log(e);
+                        }
+                        return;
+                    }
+
+                    if( fnType === 'object' ){
+                        var res = await fn;
+                        window.output = res;
+                        res = JSON.stringify(res, null, 2);
+                        result.innerText = res;
+                        console.log(res);
+                        return;
+                    }
+                    
+                    
+                })();">${key}</div>
+                <pre class="code">
+<code>${(typeof value === 'function' ? value : JSON.stringify(value, null, 2))}</code>
+                </pre>
+            </li>`);
+            lis = lis.join(' ');
+            var template = `
+            <div class="cat">
+                <h1>LitJsSdk_pkpClient has ${entries.length} functions</h1>
+                    <ul>
+                        ${ lis }
+                    </ul>
+                </div>
+            `;
+            root.insertAdjacentHTML('beforeend', template);
+        });
+    
+    
+
+    
+    
         if(typeof LitJsSdk_pkpCosmos === 'undefined') {
             console.error("LitJsSdk_pkpCosmos:", LitJsSdk_pkpCosmos);
         }else{
@@ -1500,7 +1574,7 @@ pre {
                     `,
                 }}
             />
-            (REACT) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-react.mjs Fri, 24 Mar 2023 17:26:23 GMT
+            (REACT) THIS FILE IS AUTOMATICALLY GENERATED FROM tools/scripts/gen-react.mjs Sat, 25 Mar 2023 03:23:38 GMT
             <div id="root"></div>
              <pre><code id="result"></code></pre>
         </>
