+++ package.json
@@ -77,7 +77,6 @@
     "dotenv": "^16.0.2",
     "dotenv-parse-variables": "^2.0.0",
     "download": "^8.0.0",
-    "eth-crypto": "^2.6.0",
     "ethers": "^5.7.1",
     "etherscan-api": "^10.2.0",
     "find-config": "^1.0.0",
@@ -93,6 +92,7 @@
     "react": "18.0.0",
     "react-dom": "18.0.0",
     "regenerator-runtime": "0.13.7",
+    "secp256k1": "^5.0.0",
     "serve": "^14.0.1",
     "start-server-and-test": "^1.14.0",
     "synthetix-js": "^2.74.1",
