+++ packages/pkp-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lit-protocol/pkp-client",
-  "version": "0.0.1",
+  "version": "2.1.95",
   "type": "commonjs",
   "license": "MIT",
   "homepage": "https://github.com/Lit-Protocol/js-sdk",
@@ -18,13 +18,13 @@
     "access": "public",
     "directory": "../../dist/packages/pkp-client"
   },
-  "main": "./dist/src/index.js",
-  "typings": "./dist/src/index.d.ts",
   "browser": {
     "crypto": false,
     "stream": false
   },
   "tags": [
     "universal"
-  ]
+  ],
+  "main": "./dist/src/index.js",
+  "typings": "./dist/src/index.d.ts"
 }
